+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-11-25
+### Changed
+- Changed `index.js` to update styles and php files in separate functions.
+- Renamed `sample-readmeConfig.json`.
+
 ## [1.1.1] - 2024-10-16
 ### Changed
 - Changed regex to include themes and plugins.